Pull SqsSourceSettings into it's own file with it's supporting classes
